Move some files to their own folder
I don't really know where i wanna put this, for now i'm gonna leave it in the project room
