+++ SIT.Manager.Avalonia/ManagedProcess/IAkiServerService.cs
@@ -2,7 +2,7 @@
 using System.Diagnostics;
 using static SIT.Manager.Avalonia.Services.AkiServerService;
 
-namespace SIT.Manager.Avalonia.Interfaces
+namespace SIT.Manager.Avalonia.ManagedProcess
 {
     public interface IAkiServerService : IManagedProcess
     {
