+++ SIT.Manager.Avalonia/ViewModels/PlayPageViewModel.cs
@@ -1,7 +1,7 @@
 ﻿using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
 using FluentAvalonia.UI.Controls;
-using SIT.Manager.Avalonia.Interfaces;
+using SIT.Manager.Avalonia.ManagedProcess;
 using SIT.Manager.Avalonia.Models;
 using System;
 using System.Collections.Generic;
@@ -59,7 +59,7 @@ namespace SIT.Manager.Avalonia.ViewModels
                 return;
             }
 
-            
+            ManagerConfig config = null;
 
             //Update config
